comp/requirements.txt
requests 
edge-tts 
imageio-ffmpeg 
pillow